api_productos/backend/productos/urls.py
from rest_framework import routers
from .views import ProductoViewSet
from rest_framework.routers import DefaultRouter
from .views import ProductoViewSet

router = DefaultRouter()
router.register(r'productos', ProductoViewSet, basename='producto')

urlpatterns = router.urls

